+++ src/app/components/ms-page/ms-page.component.html
@@ -48,7 +48,7 @@
             </div>
             -->
 
-            <or-page-viewer [pages]="pages" [firstPageUrl]="firstPageUrl" [parts]="msParts"></or-page-viewer>
+            <or-page-viewer [pages]="pages" [parts]="msParts"></or-page-viewer>
 
             <!-- MANUSCRIPT DESCRIPTION -->
