+++ src/app/components/still-image/still-image.component.ts
@@ -49,6 +49,7 @@ import { DspCompoundPosition } from '../dsp-resource';
 //import { EmitEvent, Events, UpdatedFileEventValue, ValueOperationEventService } from '../../services/value-operation-event.service';
 import { FileRepresentation } from '../file-representation';
 import { RepresentationService } from '../representation.service';
+import { Observable } from 'rxjs';
 
 
 @Component({
@@ -69,6 +70,7 @@ export class StillImageComponent implements OnChanges, OnDestroy, AfterViewInit 
   @Input() compoundNavigation?: DspCompoundPosition;
   // @Input() currentTab: string;
   // @Input() parentResource: ReadResource;
+  @Input() selectedPageNumber?: Observable<number>; 
 
   @Output() goToPage = new EventEmitter<number>();
 
@@ -188,14 +190,16 @@ export class StillImageComponent implements OnChanges, OnDestroy, AfterViewInit 
           withCredentials: true
       };
 
-      const index = this.images[0].fileValue.fileUrl.indexOf(this.images[0].fileValue.filename);
-      const pathToJson = this.images[0].fileValue.fileUrl.substring(0, index + this.images[0].fileValue.filename.length) + '/knora.json';
-
-      this._http.get(pathToJson, requestOptions).subscribe(
-          res => {
-              this.originalFilename = res['originalFilename'];
-          }
-      );
+      if (this.images && this.images.length>0) {
+        const index = this.images[0].fileValue.fileUrl.indexOf(this.images[0].fileValue.filename);
+        const pathToJson = this.images[0].fileValue.fileUrl.substring(0, index + this.images[0].fileValue.filename.length) + '/knora.json';
+  
+        this._http.get(pathToJson, requestOptions).subscribe(
+            res => {
+                this.originalFilename = res['originalFilename'];
+            }
+        );  
+      }
   }
 
   
@@ -333,6 +337,11 @@ export class StillImageComponent implements OnChanges, OnDestroy, AfterViewInit 
 
  
 
+    ngOnInit(): void {
+        if(this.selectedPageNumber) {
+            this.selectedPageNumber.subscribe(this.openPage);
+        }
+    }
