+++ src/app/models/dataviz.model.ts
@@ -1,6 +1,6 @@
+import { Page } from './page.model';
 import { Resource } from './resource.model';
 
-export interface DataViz extends Resource {
-    imageURL: string;
+export interface DataViz extends Page {
     pub: string;
 }
